pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "hcrawler"
version = "7.0.0"
description = "Authorized web discovery, attack-surface inventory, baseline, graphing, and lightweight audit CLI"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Henrique da Silva Cunha" }]
license = { text = "MIT" }
keywords = ["crawler", "security", "osint", "pentest", "recon", "cli", "web", "sarif", "audit"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Topic :: Security",
  "Topic :: Internet :: WWW/HTTP",
  "Environment :: Console",
]
dependencies = [
  "requests>=2.31.0",
  "beautifulsoup4>=4.12.0",
  "colorama>=0.4.6",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
async = ["aiohttp>=3.9.0"]
dev = ["pytest>=8.0.0", "ruff>=0.6.0", "build>=1.2.0", "twine>=5.0.0", "aiohttp>=3.9.0"]

[project.scripts]
hcrawler = "hcrawler.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 120
target-version = "py310"
